+++ CHANGELOG.md
@@ -5,8 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.3.3] - 2025-06-??
+## [2.4.0] - 2025-06-??
 
+- **SOME BREAKING CHANGES**. Modify the built-in sessions to support any kind
+  of storage for the session information. It still defaults to storing sessions
+  in cookies, but allows defining a custom `SessionStore` to store information
+  where desired.
 - Remove `charset-normalizer` dependency. This library was used only when a
   `UnicodeDecodeError` exception occurred when parsing the body of a web
   request. This can happen in two circumstances: when the client sends a